resolve comment.
