+++ test/Microsoft.Azure.SignalR.AspNet.Tests/Infrastructure/TestConnectionManager.cs
@@ -22,40 +22,5 @@ namespace Microsoft.Azure.SignalR.AspNet.Tests
             CurrentTransports.TryAdd(message.ConnectionId, transport);
             return transport;
         }
-
-        public class TestTransport : IServiceTransport
-        {
-            public long MessageCount = 0;
-            public Func<string, Task> Received { get; set; }
-            public Func<Task> Connected { get; set; }
-            public Func<Task> Reconnected { get; set; }
-            public Func<bool, Task> Disconnected { get; set; }
-            public string ConnectionId { get; set; }
-
-            public Task<string> GetGroupsToken()
-            {
-                return Task.FromResult<string>(null);
-            }
-
-            public void OnDisconnected()
-            {
-            }
-
-            public void OnReceived(string value)
-            {
-                // Only use to validate message count
-                MessageCount++;
-            }
-
-            public Task ProcessRequest(ITransportConnection connection)
-            {
-                return Task.CompletedTask;
-            }
-
-            public Task Send(object value)
-            {
-                return Task.CompletedTask;
-            }
-        }
     }
 }
